server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart

pillow


httpx


transformers
torch
torchvision
qwen_vl_utils
google-genai
accelerate
